+++ azure-pipeline.yml
@@ -39,6 +39,24 @@ variables:
   value: 'diquzart/go-app'
 
 stages:
+- stage: Test
+  displayName: Test and Coverage
+  jobs:
+  - job: Test
+    displayName: Test
+    pool:
+      vmImage: $(vmImageName)
+    steps:
+    - task: GoTool@0
+      inputs:
+        version: '1.14.1'
+
+    - task: Go@0
+      inputs:
+        command: 'test'
+        arguments: '-v ./controllers'
+        workingDirectory: '$(System.DefaultWorkingDirectory)'
+
 - stage: Build
   displayName: Build image
