--- controllers/health.go
@@ -1,13 +1,0 @@
-package controllers
-
-import (
-	"net/http"
-
-	"github.com/gin-gonic/gin"
-)
-
-func Health(ctx *gin.Context) {
-	ctx.String(
-		http.StatusOK,
-		"Working!")
-}
