+++ routers/router.go
@@ -8,11 +8,11 @@ import (
 
 func CreateTmplRender() multitemplate.Renderer {
 
-	baseTmpl := "frontend/templates/base.tmpl.html"
-	homeTmpl := "frontend/templates/home.tmpl.html"
-	aboutTmpl := "frontend/templates/about.tmpl.html"
-	nameTmpl := "frontend/templates/name.tmpl.html"
-	s404Tmpl := "frontend/templates/404.tmpl.html"
+	baseTmpl := "./public/templates/base.tmpl.html"
+	homeTmpl := "./public/templates/home.tmpl.html"
+	aboutTmpl := "./public/templates/about.tmpl.html"
+	nameTmpl := "./public/templates/name.tmpl.html"
+	s404Tmpl := "./public/templates/404.tmpl.html"
 
 	r := multitemplate.NewRenderer()
 	r.AddFromFiles("home", baseTmpl, homeTmpl)
@@ -33,14 +33,14 @@ func InitRouter() *gin.Engine {
 	r.Use(gin.Recovery())
 
 	// Frontend
-	r.Static("/static", "./frontend/static")
+	r.Static("/static", "./public/static")
 	r.HTMLRender = CreateTmplRender()
 
 	// Routes
 	r.GET("/", controllers.Home)
 	r.GET("/about", controllers.About)
 	r.GET("/api", controllers.Api)
-	r.GET("/user/:name", controllers.ApiParam)
+	r.GET("/user/:name", controllers.UrlParam)
 	r.GET("/health", controllers.Health)
 	r.NoRoute(controllers.NoFound)
